+++ R/ingest.R
@@ -30,7 +30,7 @@ ingest_student_data_extract <- function(path, clean_test_names = TRUE, ...) {
     # ensure these are coded as numeric for later functions
     num_vars <- c("test_scaled_score", "race", "performance_level", "disability_status", "el_overall_proficiency_level")
 
-    tmp <- dplyr::mutate(tmp, dplyr::across(num_vars, as.numeric))
+    tmp <- dplyr::mutate(tmp, dplyr::across(dplyr::all_of(num_vars), as.numeric))
 
     ret <- if (clean_test_names) {
         tmp |>