+++ R/pass_rates.R
@@ -1,10 +1,12 @@
-#' Calculate Pass Rates
+#' Summarize Pass Rates
 #'
 #' @description Calculate pass rates for a student data extract. This function applies the calculation rules described in the VDOE's annual pass rate calculation document. Retest rules are implicitly applied if the data has been preprocessed via [filter_test_performance()].
 #'
 #' @param x A dataframe containing data from an SOL student data extract. Ideally this will have been read in via [ingest_student_data_extract()] and preprocessed via [filter_test_performance()]
 #' @param group_vars NULL or a character vector specifying the column names to group the data by (e.g., "school_name", "test_name"). If `NULL` (the default), no grouping will occur.
 #' @param drop_parent_requested Logical. If `TRUE` (the default), drop any tests that were parent requested.
+#' @param drop_failing_retests Logical. If `TRUE` (the default), exclude failing retests.
+#'
 #'
 #' @return A summarized dataframe with pass rates for each group
 #'
@@ -13,18 +15,19 @@
 #' @examples \dontrun{
 #' x <- ingest_student_data_extract("path/to/data.csv")
 #' x <- filter_test_performance(x, type = "best")
-#' pr <- summarize_pass_rate(x, group_vars = c("school_name", "test_name"))
+#' pr <- summarize_pass_rates(x, group_vars = c("school_name", "test_name"))
 #' }
 #' @md
-summarize_pass_rate <- function(x, group_vars = NULL, drop_parent_requested = TRUE) {
+summarize_pass_rates <- function(x, group_vars = NULL, drop_parent_requested = TRUE, drop_failing_retests = TRUE) {
     stopifnot(
         "`x` must be a data frrame" = is.data.frame(x),
         "`group_vars` must be a character vector" = is.character(group_vars) | is.null(group_vars),
-        "`drop_parent_requested` must be a logical" = is.logical(drop_parent_requested)
+        "`drop_parent_requested` must be a logical" = is.logical(drop_parent_requested),
+        "`drop_failing_retests` must be a logical" = is.logical(drop_failing_retests)
     )
 
     # filter out exclusions
-    tmp <- filter_exclusions(x, drop_parent_requested)
+    tmp <- filter_exclusions(x, drop_parent_requested = drop_parent_requested, drop_failing_retests = drop_failing_retests)
 
     tmp <- tmp |>
         dplyr::mutate(
@@ -57,6 +60,7 @@ summarize_pass_rate <- function(x, group_vars = NULL, drop_parent_requested = TR
 #' @param x A data frame containing student data, ideally from [ingest_student_data_extract()].
 #' @param group_vars NULL or a character vector specifying the column names to group the data by (e.g., "school_name", "test_name"). If `NULL` (the default), no grouping will occur.
 #' @param drop_parent_requested Logical. If `TRUE` (the default), exclude tests marked as parent-requested.
+#' @param drop_failing_retests Logical. If `TRUE` (the default), exclude failing retests.
 #' @param convert_performance_levels Logical. If `TRUE` (the default), convert numeric performance level codes to text labels using [recode_performance_levels()].
 #'
 #' @return A data frame summarizing performance levels.  It includes the grouping variables,
@@ -82,16 +86,17 @@ summarize_pass_rate <- function(x, group_vars = NULL, drop_parent_requested = TR
 #' )
 #' }
 #' @md
-summarize_performance_levels <- function(x, group_vars = NULL, drop_parent_requested = TRUE, convert_performance_levels = TRUE) {
+summarize_performance_levels <- function(x, group_vars = NULL, drop_parent_requested = TRUE, drop_failing_retests = TRUE, convert_performance_levels = TRUE) {
     stopifnot(
         "`x` must be a data frrame" = is.data.frame(x),
         "`group_vars` must be NULL or a character vector" = is.character(group_vars) | is.null(group_vars),
         "`drop_parent_requested` must be a logical" = is.logical(drop_parent_requested),
+        "`drop_failing_retests` must be a logical" = is.logical(drop_failing_retests),
         "`convert_performance_levels` must be a logical" = is.logical(convert_performance_levels)
     )
 
     # filter out exclusions
-    tmp <- filter_exclusions(x, drop_parent_requested)
+    tmp <- filter_exclusions(x, drop_parent_requested = drop_parent_requested, drop_failing_retests = drop_failing_retests)
 
     # convert peformance levels to text if specified
     tmp <- if (convert_performance_levels) {
@@ -120,8 +125,7 @@ summarize_performance_levels <- function(x, group_vars = NULL, drop_parent_reque
 }
 
 # utilities ----------------
-# note -- if we select the student's 'best' or 'first' tests, we don't need to apply any of the rules for retests
-filter_exclusions <- function(x, drop_parent_requested) {
+filter_exclusions <- function(x, drop_parent_requested = TRUE, drop_failing_retests = TRUE) {
     allow_lvls <- c(1, 2, 3, 4, 5, 8)
 
     tmp <- x |>
@@ -138,5 +142,11 @@ filter_exclusions <- function(x, drop_parent_requested) {
         tmp
     }
 
-    return(tmp)
+    ret <- if (drop_failing_retests) {
+        drop_failing_retests(tmp)
+    } else {
+        tmp
+    }
+
+    return(ret)
 }