+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## v 0.1.1
 
 - Add drop_irw() function to drop integrated reading and writing tests
+- Add drop_failing_retests() function